Fixed lab_2(9)

+++ prog_php/lab_2/index.php
@@ -64,9 +64,15 @@
                         break;
                     }
                 case "delete":
-                    $query = "DELETE FROM `Игроки` WHERE `№ игрока` = '$num_player' AND `Страна` = '$country'";
+                    $query = "select exists(select * from `Игроки` WHERE `№ игрока`='$num_player'  AND `Страна`='$country' AND `ФИО игрока`='$name_player' AND `Амплуа`='$amp')";
                     $result = mysqli_query($link, $query);
-                    if (!$result) echo "<h1>Невозможно удалить запись: </h1>" . mysqli_error($link);
+                    $row = mysqli_fetch_row($result);
+                    if ($row[0] > 0)
+                    {
+                        $query = "DELETE FROM `Игроки` WHERE `№ игрока` = '$num_player' AND `Страна` = '$country'";
+                        $result = mysqli_query($link, $query);
+                        if (!$result) echo "<h1>Невозможно удалить запись: </h1>" . mysqli_error($link);
+                    } else echo "<p>Запись с такими полями № игрока = $num_player, Страна = $country, ФИО игрока = $name_player, Амплуа = $amp не существует!</p>";
                     break;
                 default:
                     break;
